WEBAPP-393: Extract types to types.js
