+++ src/modules/endpoint/endpoints/cities.js
@@ -3,6 +3,7 @@
 import CityModel from '../models/CityModel'
 import { apiUrl } from '../constants'
 import EndpointBuilder from '../EndpointBuilder'
+import type { JsonCityType } from '../types'
 
 const stripSlashes = (path: string): string => {
   if (path.startsWith('/')) {
@@ -12,15 +13,6 @@ const stripSlashes = (path: string): string => {
     path = path.substr(0, path.length - 1)
   }
   return path
-}
-
-type JsonCityType = {
-  name: string,
-  path: string,
-  live: boolean,
-  events: boolean,
-  extras: boolean,
-  name_without_prefix: string
 }
 
 export default new EndpointBuilder<void, Array<CityModel>>('cities')
