+++ src/modules/endpoint/endpoints/extras.js
@@ -5,6 +5,7 @@ import ExtraModel from '../models/ExtraModel'
 import EndpointBuilder from '../EndpointBuilder'
 
 import ParamMissingError from '../errors/ParamMissingError'
+import type { JsonExtraPostType, JsonExtraType } from '../types'
 
 const EXTRAS_ENDPOINT_NAME = 'extras'
 
@@ -12,18 +13,6 @@ const createPostMap = (jsonPost: JsonExtraPostType): Map<string, string> => {
   const map = new Map()
   Object.keys(jsonPost).forEach(key => map.set(key, jsonPost[key]))
   return map
-}
-
-type JsonExtraPostType = {
-  [key: string]: string
-}
-
-type JsonExtraType = {
-  alias: string,
-  name: string,
-  url: string,
-  thumbnail: string,
-  post: ?JsonExtraPostType
 }
 
 type ParamsType = { city: ?string, language: ?string }
