+++ src/modules/endpoint/endpoints/languages.js
@@ -4,12 +4,11 @@ import LanguageModel from '../models/LanguageModel'
 import { apiUrl } from '../constants'
 import EndpointBuilder from '../EndpointBuilder'
 import ParamMissingError from '../errors/ParamMissingError'
+import type { JsonLanguageType } from '../types'
 
 const LANGUAGES_ENDPOINT_NAME = 'languages'
 
 type ParamsType = { city: ?string }
-
-type JsonLanguageType = { code: string, native_name: string }
 
 export default new EndpointBuilder<ParamsType, Array<LanguageModel>>(LANGUAGES_ENDPOINT_NAME)
   .withParamsToUrlMapper((params): string => {
