+++ src/modules/endpoint/endpoints/sprungbrettJobs.js
@@ -8,10 +8,6 @@ const SPRUNGBRETT_JOBS_ENDPOINT_NAME = 'sprungbrettJobs'
 
 type ParamsType = { url: ?string }
 
-type JsonSprungbrettJobType = {
-  title: string, zip: string, city: string, url: string, employment: string, apprenticeship: string
-}
-
 export default new EndpointBuilder<ParamsType, Array<SprungbrettJobModel>>(SPRUNGBRETT_JOBS_ENDPOINT_NAME)
   .withParamsToUrlMapper((params): string => {
     if (!params.url) {
